refactor: presentation.model.quiz 패키지 생성 후 파일 이동
