--- presentation/src/main/java/com/paykidscompose/presentation/model/QuizClearedUIModel.kt
@@ -1,6 +1,0 @@
-package com.paykidscompose.presentation.model
-
-data class QuizClearedUIModel(
-    val message: String,
-    val isCleared: Boolean
-): UIModel()
