+++ presentation/src/main/java/com/paykidscompose/presentation/model/quiz/QuizClearedUIModel.kt
@@ -1,0 +1,8 @@
+package com.paykidscompose.presentation.model.quiz
+
+import com.paykidscompose.presentation.model.UIModel
+
+data class QuizClearedUIModel(
+    val message: String,
+    val isCleared: Boolean
+): UIModel()
